Move KeywordUpsell to modals folder
